requirements.txt
bittensor==9.10.0
psutil>=5.9.0
cryptography>=41.0.0
pydantic>=2.0.0
py3nvml>=0.2.7
PyYAML>=6.0.0
py-cpuinfo>=9.0.0
websockets>=11.0.0
loguru>=0.7.2
